requirements.txt
# Web framework and ASGI server – core
fastapi==0.109.1
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
Jinja2==3.1.3
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
itsdangerous==2.2.0

# HTTP and API clients
requests==2.31.0
httpx==0.27.0
cachetools==5.3.3

# Modern UI framework
nicegui==1.4.21
reactpy==1.1.0

# Production server
gunicorn==21.2.0

# Async HTTP client
aiohttp==3.9.3
